Add files via upload
Build a multimodal Gaussian distribution with synthetic data.  
Construct a piecewise linear regression and plot the result with its splines 
and knots. 
Construct a piecewise polynomial regression and plot the result, its 
splines, and knots. 
In addition to F-statistics that you use to decide about the number of 
knots, report RSME and R^2 for all models.  
Try to model the multimodal Gaussian distribution built with single 
polynomial regression. 
Measure and report the execution time of tasks (2), task (3), and task 
(5).  Then report the differences in execution time (no plot required, but 
you need to report them in a table). Hint: use excel for table drawing and 
add it in the word file. 
Use the Multimodal Gaussian distribution of Tasks 5, apply Ridge, 
LASSO and compare their accuracy and number of parameters, 
parameters coefficient with Polynomial regression. Here you should also 
report and discuss the differences (if there are any differences)